# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRKR Invoice Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvoiceBody {
  invoiceNumber: string;
  invoiceAmount: number;
  invoiceDate?: string | null;
  invoiceType?: string | null;
  fiscalYear?: string | null;
  vendorID?: string | null;
  vendorName?: string | null;
  contractPONumber?: string | null;
  fiscalPONumber?: string | null;
  reportingStructure?: string | null;
  submissionNotes?: string | null;
  programReceivedDate?: string | null;
  accountingReceivedDate?: string | null;
  processingType?: string | null;
  expedite?: boolean;
  cashHold?: boolean;
  localHealth?: boolean;
  specialHandling?: boolean;
  dvbeSbCmia?: boolean;
  revolvingFund?: boolean;
  snapNurse?: boolean;
  calVaxGrant?: boolean;
  schoolsGrant?: boolean;
  airQualityCmp?: boolean;
  advancePayment?: boolean;
  drill2?: boolean;
  covid19Related?: string | null;
  submitterName?: string | null;
  submitterEmail?: string | null;
}