# lib/api-zod/src/generated/types/purchaseOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRKR Invoice Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseOrder {
  id: number;
  poNumber: string;
  poDate?: string | null;
  poStatus?: string | null;
  supplierID?: string | null;
  supplierName?: string | null;
  lineItemDescription?: string | null;
  fund?: string | null;
  program?: string | null;
  project?: string | null;
  encumberedAmount?: number | null;
  expensedAmount?: number | null;
  remainingEncumbrance?: number | null;
  budgetStatus?: string | null;
  createdAt: string;
}